Move tests to the root
Closes #69.
